requirements.txt
streamlit>=1.20.0
transformers>=4.24.0
deepface>=0.0.75
opencv-python-headless>=4.5.5.64
numpy>=1.21.2
Pillow>=8.2.0
pandas>=1.3.3
matplotlib>=3.4.3
plotly>=5.0.0
sqlite3
